+++ R/00_global_variables.R
@@ -1,5 +1,6 @@
 globalVariables(
   names = c(
-    "walk_number","x","y","value","name", "cum_sum","cum_prod",".tooltip"
+    "walk_number","x","y","value","name", "cum_sum","cum_prod",".tooltip",
+    "step_number",""
   )
 )